requirements.txt
streamlit==1.37.0
transformers==4.43.3
torch==2.3.1
pillow==10.4.0
huggingface_hub==0.24.2

